# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
  # local:
  - url: "http://127.0.0.1:3000"
    # remote:
    # - url: "http://132.72.65.211"
    description: SwaggerHub API Auto Mocking hey
...
paths:
  /Register:
    post:
      tags:
      - auth
      summary: Create a new user account
      description: Register a new user to the system
      operationId: createUser
      requestBody:
        description: User to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: user created successfully
        "400":
          description: "invalid input, object invalid"
        "409":
          description: user already exist
  /Login:
    post:
      tags:
      - auth
      summary: Log in a user
      description: Authenticate user and return session cookie
      operationId: loginUser
      requestBody:
        description: User credentials for login
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginScheme'
        required: true
      responses:
        "200":
          description: |
            User logged in successfully. The session ID is returned in a cookie named `session`.
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: invalid username or password
...
  /recipes/search:
    get:
      tags:
      - receipe
      summary: get receipes from the server
      description: "By passing in the appropriate options, you can get \navailable receipes in the system\n"
      operationId: getRecipeByText
      parameters:
      - name: number
        in: query
        description: Number of desired recipes to return
        required: true
        schema:
          type: integer
          default: 5
          example: 5
          enum:
          - 5
          - 10
          - 15
      - name: query
        in: query
        description: Text to search for matching recipes
        required: true
        schema:
          type: string
          example: pasta
      - name: cuisine
        in: query
        description: Filter recipes by one or more cuisine types (comma-separated)
        required: false
        style: form
        explode: false
        schema:
          type: array
          example:
          - Italian
          - Greek
          items:
            type: string
            enum:
            - African
            - Asian
            - American
            - British
            - Cajun
            - Caribbean
            - Chinese
            - Eastern European
            - European
            - French
            - German
            - Greek
            - Indian
            - Irish
            - Italian
            - Japanese
            - Jewish
            - Korean
            - Latin American
            - Mediterranean
            - Mexican
            - Middle Eastern
            - Nordic
            - Southern
            - Spanish
            - Thai
            - Vietnamese
      - name: diet
        in: query
        description: Filter recipes by one or more diet types (comma-separated)
        required: false
        style: form
        explode: false
        schema:
          type: array
          example:
          - vegetarian
          - vegan
          items:
            type: string
            enum:
            - gluten free
            - ketogenic
            - vegetarian
            - lacto-vegetarian
            - ovo-vegetarian
            - vegan
            - pescetarian
            - paleo
            - primal
            - low FODMAP
            - whole30
      - name: intolerances
        in: query
        description: Filter recipes by one or more intolerances (comma-separated)
        required: false
        style: form
        explode: false
        schema:
          type: array
          example:
          - gluten
          - peanut
          - dairy
          items:
            type: string
            enum:
            - dairy
            - egg
            - gluten
            - grain
            - peanut
            - seafood
            - sesame
            - shellfish
            - soy
            - sulfite
            - tree nut
            - wheat
      responses:
        "200":
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Receipe'
        "400":
          description: bad input parameter
...
  /recipes/like:
    get:
      tags:
        - receipe
      summary: Get number of likes for a recipe
      description: Retrieve the number of likes for a specific recipe.
      operationId: getRecipeLikes
      parameters:
        - name: recipeId
          in: query
          required: true
          schema:
            type: integer
          description: The ID of the recipe
      responses:
        "200":
          description: Number of likes retrieved successfully
          content:
            application/json:
              schema:
                type: integer
                example: 42
        "400":
          description: Invalid recipe ID
...
  /recipes/random:
    get:
      tags:
      - receipe
      summary: Get random recipes
      description: Get a random selection of recipes from the external API (Spoonacular)
      operationId: getRandomRecipes
      parameters:
      - name: number
        in: query
        description: Number of random recipes to return
        required: false
        schema:
          type: integer
          default: 3
          example: 3
      responses:
        "200":
          description: Random recipes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Receipe'
        "401":
          description: internal server error
        "500":
          description: External API error or server error
  /recipes/searchById/{recipeId}:
    get:
      tags:
      - receipe
      summary: Get detailed recipe information
      description: Get comprehensive information about a specific recipe from external API (Spoonacular)
      operationId: getRecipeInformation
      parameters:
      - name: recipeId
        in: path
        description: The ID of the recipe to get information for
        required: true
        schema:
          type: integer
          example: 716429
      - name: includeNutrition
        in: query
        description: Whether to include nutrition information
        required: false
        schema:
          type: boolean
          default: false
          example: false
      responses:
        "200":
          description: Detailed recipe information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipe'
        "400":
          description: Invalid recipe ID format
        "404":
          description: Recipe not found
        "500":
          description: External API error or server error
...
components:
  schemas:
    Receipe:
      required:
      - summary
      - glutenFree
      - image
      - extendedIngredients
      - analyzedInstructions
      - title
      - readyInMinutes
      - servings
      - vegan
      - sourceName
      - vegetarian
      type: object
      properties:
        id:
          type: integer
          example: 10
        title:
          type: string
          example: Spaghetti Bolognese
        image:
          type: string
          format: URL
          example: https://example.com/images/spaghetti.jpg
        readyInMinutes:
          type: integer
          example: 30
        servings:
          type: integer
          example: 4
        vegetarian:
          type: boolean
          example: false
        vegan:
          type: boolean
          example: false
        glutenFree:
          type: string
          example: false
        summary:
          type: string
          example: A classic Italian pasta dish with a rich meat sauce.
        extendedIngredients:
          type: array
          description: Detailed list of ingredients with metadata, quantities, and units
          items:
            type: object
            properties:
              id:
                type: integer
                example: 12061
              aisle:
                type: string
                example: "Nuts"
              image:
                type: string
                example: "almonds.jpg"
              consistency:
                type: string
                example: "SOLID"
              name:
                type: string
                example: "almond cookies"
              nameClean:
                type: string
                example: "almond cookies"
              original:
                type: string
                example: "150 grams Almond cookies, crumbed"
              originalName:
                type: string
                example: "Almond cookies, crumbed"
              amount:
                type: number
                example: 150.0
              unit:
                type: string
                example: "grams"
              meta:
                type: array
                items:
                  type: string
                example: []
              measures:
                type: object
                properties:
                  us:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 5.291
                      unitShort:
                        type: string
                        example: "oz"
                      unitLong:
                        type: string
                        example: "ounces"
                  metric:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 150.0
                      unitShort:
                        type: string
                        example: "g"
                      unitLong:
                        type: string
                        example: "grams"
        analyzedInstructions:
          type: array
          description: "Structured step-by-step preparation instructions"
          items:
            type: object
            properties:
              name:
                type: string
              steps:
                type: array
                items:
                  type: object
                  properties:
                    number:
                      type: integer
                    step:
                      type: string
                    ingredients:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          localizedName:
                            type: string
                          image:
                            type: string
                    equipment:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          localizedName:
                            type: string
                          image:
                            type: string
                          temperature:
                            type: object
                            properties:
                              number:
                                type: number
                              unit:
                                type: string
                    length:
                      type: object
                      properties:
                        number:
                          type: integer
                        unit:
                          type: string
          example:
            - name: ""
              steps:
                - number: 1
                  step: "Toss the cookie crumbs into the melted butter..."
                  ingredients:
                    - id: 10018192
                      name: "cookie crumbs"
                      localizedName: "cookie crumbs"
                      image: ""
                    - id: 1001
                      name: "butter"
                      localizedName: "butter"
                      image: "butter-sliced.jpg"
                  equipment:
                    - id: 404628
                      name: "hand mixer"
                      localizedName: "hand mixer"
                      image: "https://spoonacular.com/cdn/equipment_100x100/hand-mixer.png"
        sourceName:
          type: string
          description: The user who created or added the recipe to the family collection.
          example: user101
        popularity:
          type: integer
          description: The number of likes the recipe has
          example: 100
    User:
      type: object
      required:
        - country
        - email
        - firstname
        - lastname
        - password
        - username
      properties:
        user_id:
          type: integer
          readOnly: true
          example: 101
        username:
          type: string
          example: Acef_
        firstname:
          type: string
          example: Asaf
        lastname:
          type: string
          example: Baroch
        email:
          type: string
          format: email
          example: asaf@example.com
        password:
          type: string
          format: password
          example: "123"
        country:
          type: string
          example: Israel
...
    LoginScheme:
      required:
      - email
      - password
      type: object
      properties:
        username:
          type: string
          example: Acef_
        password:
          type: string
          format: password
          example: "123"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session